REFACTOR: refactoring EditModal
