+++ src/modules/EditModal/index.tsx
@@ -1,6 +1,7 @@
-import React, {useState} from 'react';
+import React from 'react';
 import {View, StyleSheet, TextInput, Button, Modal, Alert} from 'react-native';
 import {THEME} from 'src/theme';
+import useEditModal from './useEditModal';
 
 type EditModalProps = {
   value: string;
@@ -15,21 +16,7 @@ const EditModal: React.FC<EditModalProps> = ({
   setVisible,
   saveItem,
 }) => {
-  const [title, setTitle] = useState<string>(value);
-
-  const saveHandler = () => {
-    if (title.trim().length < 3) {
-      Alert.alert(
-        'Error',
-        `Minimum name length is 3 characters. ${
-          title.trim().length
-        } are too few!`,
-      );
-    } else {
-      saveItem(title);
-    }
-  };
-
+  const {title, setTitle, saveHandler} = useEditModal({value, saveItem});
   return (
     <Modal
       animationType="slide"
